refactor(runner): extract runner-exec artifact promotion into core
The runner exec command owned the full artifact-promotion engine —
resolving/downloading declared runner-side outputs, recording file/
directory artifacts to the observation store, deriving typed fuzz
artifacts, and building promoted-output/structured-summary views. That
is filesystem mutation + run artifact persistence in the command layer,
tripping the thin_command_adapter audit detector.

Move the engine into core::runner::execution::artifact_promotion (next
to the runner::exec primitive it complements), operating on core runner
types. The command now runs the job via runner::exec and delegates all
promotion to module-qualified core calls. Command module drops from
1079 to 476 lines.

Also fixes a detector false positive (refs #8328): runner::exec is the
core runner primitive, so a command calling it is correct thin-adapter
delegation, not runner execution orchestration. Removed runner::exec
from the marker patterns in homeboy.json (mirrors the #8297 RunDir::create
fix). The remote_execution_preflight detector's own dispatch_markers are
unaffected.

Clears the runner/exec.rs thin_command_adapter_violation.
