+++ homeboy.json
@@ -459,7 +459,7 @@
           ],
           "label": "runner execution orchestration",
           "patterns": [
-            "\\b(runner::exec|execute_[A-Za-z0-9_]+|dispatch_[A-Za-z0-9_]+)\\s*\\("
+            "\\b(execute_[A-Za-z0-9_]+|dispatch_[A-Za-z0-9_]+)\\s*\\("
           ]
         }
       ]
