+++ src/commands/runner/tests/exec.rs
@@ -3,13 +3,15 @@ use super::super::dispatch::{
 };
 use super::super::exec::{
     exec, exec_workspace_context, prepare_runner_exec_command, prepare_runner_exec_env,
-    prepare_runner_exec_secret_env_plan, promote_runner_exec_artifact_dirs,
-    promote_runner_exec_artifacts, read_bounded, read_runner_exec_script,
+    prepare_runner_exec_secret_env_plan, read_bounded, read_runner_exec_script,
     validate_runner_exec_public_env, RUNNER_EXEC_SCRIPT_LIMIT_BYTES,
 };
 
 use homeboy::core::observation::{NewRunRecord, ObservationStore};
-use homeboy::core::runners::{self as runner, RunnerExecMode, RunnerExecOutput};
+use homeboy::core::runners::{
+    self as runner, promote_runner_exec_artifact_dirs, promote_runner_exec_artifacts,
+    RunnerExecMode, RunnerExecOutput,
+};
 use homeboy::core::server;
 
 #[test]
